+++ EventParams.cpp
@@ -10,7 +10,7 @@ bool EventParams::Find(const std::string& szParamName) const
   return m_params.find(szParamName) != m_params.end();
 }
 
-int EventParams::Count() const
+size_t EventParams::Count() const
 {
 	return m_params.size();
 }
